enh: added support for LongTinyType; fix: {Byte|Short}.toString instead of Integer.toString

+++ examples/src/main/java/org/anima/tinytypes/jersey/ByteTTHeaderDelegateProvider.java
@@ -21,7 +21,7 @@ public class ByteTTHeaderDelegateProvider<T extends ByteTinyType> implements Hea
 
     @Override
     public String toString(ByteTinyType value) {
-        return Integer.toString(value.value);
+        return Byte.toString(value.value);
     }
 
     @Override
